# docs/swagger.yaml
definitions:
  app.Pager:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_rows:
        type: integer
    type: object
  errcode.Error:
    type: object
  model.Article:
    properties:
      content:
        type: string
      cover_image_url:
        type: string
      created_by:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      desc:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      state:
        type: integer
      title:
        type: string
    type: object
  model.ArticleSwagger:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Article'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
    type: object
  model.Tag:
    properties:
      created_by:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      id:
        type: integer
      is_del:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      name:
        type: string
      state:
        type: integer
    type: object
  model.TagSwagger:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
    type: object
info:
  contact: {}
  description: A blog backend program developed with Gin.
  title: blogie
  version: "1.0"
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: article name
        in: query
        name: name
        type: string
      - description: tag id
        in: query
        name: tag_id
        type: integer
      - description: state
        in: query
        name: state
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.ArticleSwagger'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Get Article List
    post:
      parameters:
      - description: tag id
        in: body
        name: tag_id
        required: true
        schema:
          type: string
      - description: article name
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: description
        in: body
        name: desc
        schema:
          type: string
      - description: cover image url
        in: body
        name: cover_image_url
        required: true
        schema:
          type: string
      - description: article content
        in: body
        name: content
        required: true
        schema:
          type: string
      - description: creator
        in: body
        name: created_by
        required: true
        schema:
          type: string
      - description: state
        in: body
        name: state
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Add A New Article
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Delete An Article
    get:
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Get An Article
    put:
      parameters:
      - description: tag id
        in: body
        name: tag_id
        schema:
          type: string
      - description: article name
        in: body
        name: title
        schema:
          type: string
      - description: description
        in: body
        name: desc
        schema:
          type: string
      - description: cover image url
        in: body
        name: cover_image_url
        required: true
        schema:
          type: string
      - description: article content
        in: body
        name: content
        required: true
        schema:
          type: string
      - description: updator
        in: body
        name: modified_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Update An Article
  /api/v1/tags:
    get:
      parameters:
      - description: tag name
        in: query
        maxLength: 100
        name: name
        type: string
      - default: 1
        description: state
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Get Tag List
    post:
      parameters:
      - description: tag name
        in: body
        maxLength: 100
        minLength: 3
        name: name
        required: true
        schema:
          type: string
      - default: 1
        description: state
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      - description: creator
        in: body
        maxLength: 100
        minLength: 3
        name: created_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Add A New Tag
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: tag id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Delete A Tag
    get:
      parameters:
      - description: tag id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.TagSwagger'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Get A Tag
    put:
      parameters:
      - description: tag id
        in: path
        name: id
        required: true
        type: integer
      - description: tag name
        in: body
        maxLength: 100
        minLength: 3
        name: name
        schema:
          type: string
      - default: 1
        description: state
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      - description: updator
        in: body
        maxLength: 100
        minLength: 3
        name: modified_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Update A Tag
swagger: "2.0"